+++ test/onyx/triggers/trigger_watermark_test.clj
@@ -1,6 +1,7 @@
-(ns onyx.windowing.trigger-watermark-test
+(ns onyx.triggers.trigger-watermark-test
   (:require [clojure.test :refer [deftest is]]
-            [onyx.triggers.triggers-api :as api]
+            [onyx.peer.window-state :refer [map->StateEvent]]
+            [onyx.triggers]
             [onyx.api]))
 
 (deftest watermark-test
@@ -12,14 +13,15 @@
                 :window/window-key :event-time
                 :window/range [5 :minutes]}
         trigger {:trigger/window-id :collect-segments
-                 :trigger/refinement :accumulating
+                 :trigger/refinement :onyx.triggers.refinements/accumulating
                  :trigger/on :watermark
                  :trigger/sync ::no-op
                  :trigger/id :trigger-id}
         segment {:event-time t}
         event {}]
-    (is (api/trigger-fire? event trigger {:window window :segment segment
-                                          :upper-extent (dec t)})))
+    (is (onyx.triggers/watermark-fire? trigger nil (map->StateEvent {:window window 
+                                                                     :segment segment
+                                                                     :upper-bound (dec t)}))))
 
   (let [t (System/currentTimeMillis)
         window {:window/id :collect-segments
@@ -29,7 +31,7 @@
                 :window/window-key :event-time
                 :window/range [5 :minutes]}
         trigger {:trigger/window-id :collect-segments
-                 :trigger/refinement :accumulating
+                 :trigger/refinement :onyx.triggers.refinements/accumulating
                  :trigger/on :watermark
                  :trigger/sync ::no-op
                  :trigger/id :trigger-id}
@@ -37,5 +39,6 @@
         event {}]
     (is
      (not
-      (api/trigger-fire? event trigger {:window window :segment segment
-                                        :upper-extent (inc t)})))))
+       (onyx.triggers/watermark-fire? trigger nil (map->StateEvent {:window window 
+                                                                    :segment segment
+                                                                    :upper-bound (inc t)}))))))