+++ test/onyx/windowing/aggregation_min_test.clj
@@ -28,8 +28,9 @@
 
 (def test-state (atom []))
 
-(defn update-atom! [event window trigger {:keys [window-id upper-bound lower-bound]} state]
-  (swap! test-state conj [lower-bound upper-bound state]))
+(defn update-atom! [event window trigger {:keys [lower-bound upper-bound event-type] :as opts} extent-state]
+  (when-not (= :task-lifecycle-stopped event-type)
+    (swap! test-state conj [lower-bound upper-bound extent-state])))
 
 (def in-chan (atom nil))
 
@@ -90,9 +91,8 @@
 
         triggers
         [{:trigger/window-id :collect-segments
-          :trigger/refinement :accumulating
-          :trigger/on :segment
-          :trigger/fire-all-extents? true
+          :trigger/refinement :onyx.triggers.refinements/accumulating
+          :trigger/on :onyx.triggers/segment
           :trigger/threshold [15 :elements]
           :trigger/sync ::update-atom!}]