+++ test/onyx/windowing/window_session_test.clj
@@ -60,11 +60,11 @@
 
 (def test-state (atom []))
 
-(defn update-atom! [event window trigger {:keys [window-id upper-bound lower-bound]} state]
-  (swap! test-state conj
-         [(java.util.Date. lower-bound)
-          (java.util.Date. upper-bound)
-          state]))
+(defn update-atom! [event window trigger {:keys [lower-bound upper-bound event-type] :as opts} extent-state]
+  (when-not (= :task-lifecycle-stopped event-type)
+    (swap! test-state conj [(java.util.Date. lower-bound)
+                            (java.util.Date. upper-bound)
+                            extent-state])))
 
 (def in-chan (atom nil))
 
@@ -126,9 +126,9 @@
 
         triggers
         [{:trigger/window-id :collect-segments
-          :trigger/refinement :accumulating
-          :trigger/on :segment
+          :trigger/refinement :onyx.triggers.refinements/accumulating
           :trigger/fire-all-extents? true
+          :trigger/on :onyx.triggers/segment
           :trigger/threshold [15 :elements]
           :trigger/sync ::update-atom!}]